Improve generalization variant support of all tasks
Generalization variant information is now stored for all executed
tasks (cross-variant tasks have a variant value of `null`) and
used for task prioritization decisions of the task queue.

Furthermore, logs of executed `ConsoleCommand`s are now preserved
for all variants (previously, logs from earlier variants were
overwritten by logs from later variants).

+++ src/main/java/teralizer/processing/task/CleanupTask.java
@@ -1,9 +1,11 @@
 package teralizer.processing.task;
 
 import org.apache.commons.io.FileUtils;
+import org.jooq.generated.tables.records.ProjectRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import teralizer.TestGeneralizationRunner;
+import teralizer.processing.GeneralizationVariant;
 import teralizer.processing.ProcessingStage;
 import teralizer.processing.TaskContext;
 
@@ -32,8 +34,19 @@ public class CleanupTask extends AbstractTask {
         this.testSourcePath = testSourcePath;
     }
 
+    public CleanupTask(ProcessingStage stage, GeneralizationVariant variant, ProjectRecord projectRecord) {
+        this.stage = stage;
+        this.variant = variant;
+        this.projectRecord = projectRecord;
+        this.projectPath = projectRecord.getRootPath();
+        this.testSourcePath = projectRecord.getTestSourcePath();
+    }
+
     @Override
     protected void executeInternal(TaskContext context, Consumer<String> reportInfo, Consumer<Task> scheduleTask) throws Exception {
+        if (this.stage != ProcessingStage.CLEANUP && this.stage != ProcessingStage.CLEANUP_GENERALIZATION) {
+            throw new RuntimeException("Cannot preform cleanup. Unsupported processing stage " + this.stage + ".");
+        }
         if (this.projectPath == null) {
             LOGGER.atInfo().log("Skipping cleanup. Project path is null.");
             return;
@@ -47,14 +60,16 @@ public class CleanupTask extends AbstractTask {
         // project. Even though this could be figured out, deleting all files that we
         // might have created in previous runs for any project type is much easier...
 
-        File mavenBuildFile = this.projectPath.resolve(ProjectSetupTask.MAVEN_CUSTOM_BUILD_FILE).toFile();
-        File gradleBuildFile =  this.projectPath.resolve(ProjectSetupTask.GRADLE_CUSTOM_BUILD_FILE).toFile();
-        List<File> buildFiles = Arrays.asList(mavenBuildFile, gradleBuildFile);
-        for (File buildFile : buildFiles) {
-            if (buildFile.exists()) {
-                LOGGER.atWarn().log("Deleting build file '" + buildFile + "'.");
-                if (!buildFile.delete()) {
-                    throw new RuntimeException("Failed to delete build file '" + buildFile + "'.");
+        if (this.stage == ProcessingStage.CLEANUP) {
+            File mavenBuildFile = this.projectPath.resolve(ProjectSetupTask.MAVEN_CUSTOM_BUILD_FILE).toFile();
+            File gradleBuildFile =  this.projectPath.resolve(ProjectSetupTask.GRADLE_CUSTOM_BUILD_FILE).toFile();
+            List<File> buildFiles = Arrays.asList(mavenBuildFile, gradleBuildFile);
+            for (File buildFile : buildFiles) {
+                if (buildFile.exists()) {
+                    LOGGER.atWarn().log("Deleting build file '" + buildFile + "'.");
+                    if (!buildFile.delete()) {
+                        throw new RuntimeException("Failed to delete build file '" + buildFile + "'.");
+                    }
                 }
             }
         }
@@ -67,31 +82,66 @@ public class CleanupTask extends AbstractTask {
             testSourcePath = this.projectPath;
         }
 
-        String dirToDelete = TestGeneralizationRunner.TOOL_NAME.toLowerCase() + "_generated";
-        Files.walkFileTree(testSourcePath, new SimpleFileVisitor<Path>() {
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-                if (dir.getFileName().toString().equals(dirToDelete)) {
-                    LOGGER.atInfo().log("Deleting directory '" + dir + "'.");
-                    FileUtils.deleteDirectory(dir.toFile());
-                    return FileVisitResult.SKIP_SUBTREE;
-                }
-                return FileVisitResult.CONTINUE;
-            }
-        });
+        Files.walkFileTree(testSourcePath, new DirectoryDeletionVisitor(this.stage, this.variant));
 
         // We do not automatically remove collected data in the DB and the data
         // directory. These should be preserved even if the generalization is
         // reverted to enable comparisons across multiple generalization runs.
     }
 
+    private static class DirectoryDeletionVisitor extends SimpleFileVisitor<Path> {
+
+        private final ProcessingStage stage;
+        private final GeneralizationVariant variant;
+
+        public DirectoryDeletionVisitor(ProcessingStage stage, GeneralizationVariant variant) {
+            this.stage = stage;
+            this.variant = variant;
+        }
+
+        @Override
+        public FileVisitResult preVisitDirectory(Path directory, BasicFileAttributes attributes) throws IOException {
+            if (this.shouldDeleteDirectory(directory)) {
+                LOGGER.atInfo().log("Deleting directory '" + directory + "'.");
+                FileUtils.deleteDirectory(directory.toFile());
+                return FileVisitResult.SKIP_SUBTREE;
+            }
+            return FileVisitResult.CONTINUE;
+        }
+
+        private boolean shouldDeleteDirectory(Path directory) {
+            if (this.stage == ProcessingStage.CLEANUP) {
+                return this.isGeneratedDirectory(directory);
+            } else if (this.stage == ProcessingStage.CLEANUP_GENERALIZATION) {
+                return this.isGeneratedDirectory(directory.getParent()) && this.isVariantDirectory(directory);
+            } else {
+                throw new RuntimeException("Unsupported processing stage " + this.stage + ".");
+            }
+        }
+
+        private boolean isGeneratedDirectory(Path directory) {
+            return directory.getFileName().toString().equals(TestGeneralizationRunner.TOOL_NAME.toLowerCase() + "_generated");
+        }
+
+        private boolean isVariantDirectory(Path directory) {
+            for (GeneralizationVariant variant : GeneralizationVariant.values()) {
+                if (directory.getFileName().toString().equals(variant.toString().toLowerCase())) {
+                    return true;
+                }
+            }
+            return false;
+        }
+    }
+
     @Override
     public String toString() {
-        return "CleanupTask{" +
-            "stage=" + this.stage +
-            ", projectPath=" + this.projectPath +
-            ", testSourcePath=" + this.testSourcePath +
-            '}';
+        String str = this.getClass().getSimpleName() + "{";
+        str += "stage=" + this.getStage();
+        str += this.getVariant() == null ? "" : ", tool=" + this.getVariant();
+        str += ", projectPath=" + this.projectPath;
+        str += ", testSourcePath=" + this.testSourcePath;
+        str += "}";
+        return str;
     }
 
     @Override
@@ -99,11 +149,14 @@ public class CleanupTask extends AbstractTask {
         if (this == o) return true;
         if (!(o instanceof CleanupTask)) return false;
         CleanupTask that = (CleanupTask) o;
-        return this.stage == that.stage && Objects.equals(this.projectPath, that.projectPath) && Objects.equals(this.testSourcePath, that.testSourcePath);
+        return this.stage == that.stage
+            && this.variant == that.variant
+            && Objects.equals(this.projectPath, that.projectPath)
+            && Objects.equals(this.testSourcePath, that.testSourcePath);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(this.stage, this.projectPath, this.testSourcePath);
+        return Objects.hash(this.stage, this.variant, this.projectPath, this.testSourcePath);
     }
 }
